requirements.txt
PyQt5
SQLAlchemy
python-dotenv
python-barcode
Pillow
reportlab
